+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v2.6.1-SNAPSHOT
 * Add support for updating user profile custom fields to ID
+* Fix using local storage for ShowRecommendationsController
 * Add other Composer external events
 
 ## v2.6.0